+++ tests/testthat/test-vlmc_simulate.R
@@ -61,3 +61,35 @@ test_that("vlmc simulate respects the type of the data", {
     expect_type(simulate(d_model, 50), typeof(data))
   }
 })
+
+test_that("vlmc simulate handles correctly burnin", {
+  for (k in 2:4) {
+    data_set <- build_markov_chain(1000, k, seed = k)
+    x_vlmc <- vlmc(data_set$x, alpha = 0.05)
+    xs <- simulate(x_vlmc, nsim = 250, seed = 5, burnin = 50)
+    expect_length(xs, 250)
+    xs_full <- simulate(x_vlmc, nsim = 300, seed = 5, burnin = 0)
+    expect_length(xs_full, 300)
+    expect_equal(xs, xs_full[-(1:50)])
+    xs_auto <- simulate(x_vlmc, nsim = 250, seed = 10, burnin = "auto")
+    xs_auto_full <- simulate(x_vlmc, nsim = 250 + 64 * context_number(x_vlmc), seed = 10, burnin = 0)
+    expect_equal(
+      xs_auto,
+      xs_auto_full[(length(xs_auto_full) - 249):length(xs_auto_full)]
+    )
+  }
+})
+
+test_that("vlmc simulate handles correctly burnin with init values", {
+  for (k in 2:4) {
+    data_set <- build_markov_chain(1000, k, seed = k)
+    x_vlmc <- vlmc(data_set$x, alpha = 0.05)
+    xs_init <- sample(states(x_vlmc), 25, replace = TRUE)
+    xs <- simulate(x_vlmc, nsim = 250, seed = 5, burnin = 50, init = xs_init)
+    expect_length(xs, 250)
+    xs_full <- simulate(x_vlmc, nsim = 300, seed = 5, burnin = 0, init = xs_init)
+    expect_length(xs_full, 300)
+    expect_equal(xs, xs_full[-(1:50)])
+    expect_equal(xs_full[1:25], xs_init)
+  }
+})