+++ NEWS.md
@@ -14,6 +14,7 @@
 * `cutoff` uses a new `tolerance` parameter to avoid reporting cut off values
   that are almost identical due to numerical imprecision 
 * `trim.covlmc` implements simple trimming for VGAM based objects (issue #48)
+* `simulate.vlmc` implements a user specified burn in period (issue #40)
   
 ## Minor improvements and bug fixes
 * Improved the test coverage (issue #54)